+++ src/Repository/Query.php
@@ -10,6 +10,7 @@ use Nette\Database\Table\Selection;
 class Query extends Selection implements QueryInterface
 {
     use QueryTrait;
+    use QuerySelectionTrait;
 
     public const CHUNK_SIZE = 127;